@@ -4,10 +4,10 @@ version: 1.0.0
 homepage:
 
 environment:
-  sdk: ">=2.18.0 <3.0.0"
+  sdk: '>=3.0.0 <4.0.0'
 
 dependencies:
 
 dev_dependencies:
-  test: "^1.19.3"
-  flutter_lints: "^1.0.4"
+  test: '^1.19.3'
+  flutter_lints: '^1.0.4'
