@@ -3,8 +3,8 @@ description: helpful widgets
 version: 0.0.1
 
 environment:
-  sdk: '>=2.18.2 <3.0.0'
-  flutter: ">=1.17.0"
+  sdk: '>=3.0.0 <4.0.0'
+  flutter: '>=1.17.0'
 
 dependencies:
   flutter:
