@@ -6,7 +6,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.18.0 <3.0.0"
+  sdk: '>=3.0.0 <4.0.0'
 
 dependencies:
   flutter:
@@ -14,11 +14,9 @@ dependencies:
   provider: ^6.0.3
   reorderables:
   flutter_reified_lenses:
-    path:
-      ../flutter_reified_lenses
+    path: ../flutter_reified_lenses
   reified_lenses:
-    path:
-      ../reified_lenses
+    path: ../reified_lenses
   flutter_hooks: ^0.18.5
   flutter_neumorphic: ^3.1.0
   path_provider: ^2.0.0
