@@ -5,7 +5,7 @@ publish_to: none
 # homepage: https://www.example.com
 
 environment:
-  sdk: '>=2.18.0 <3.0.0'
+  sdk: '>=3.0.0 <4.0.0'
 
 dependencies:
   parse_generate:
@@ -17,5 +17,5 @@ dependencies:
 
 dev_dependencies:
   lints: ^2.0.0
-  test: 
+  test:
   build_runner: ^2.3.0
