@@ -3,9 +3,8 @@ version: 0.1.0
 description: >-
   Code generation for reified lenses.
 publish_to: none
-
 environment:
-  sdk: ">=2.18.0 <3.0.0"
+  sdk: '>=3.0.0 <4.0.0'
 
 dependencies:
   meta: ^1.3.0-nullsafety.6
